File restructure and code cleanup
